Updated Serializers, Views

+++ movielix/views.py
@@ -3,7 +3,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from .models import Movie, Tag, Collection, Genre, Watchlist, MovieReview, Favorite
-from .serializers import MovieSerializer, TagSerializer, CollectionSerializer, GenreSerializer, WatchlistSerializer, MovieReviewSerializer, FavoriteSerializer
+from .serializers import MovieSerializer, TagSerializer, CollectionSerializer, GenreSerializer, WatchlistSerializer, MovieReviewSerializer, FavoriteSerializer, CollectionCreateSerializer
 from django.contrib.auth.password_validation import validate_password
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from django.core.exceptions import ValidationError
@@ -47,6 +47,8 @@ class MovieDetailView(APIView):
 
     def patch(self, request, pk):
         movie =self.get_object(pk)
+        if movie.user != request.user:
+            return Response({"error": "You do not have permission to edit this movie."}, status=status.HTTP_403_FORBIDDEN)
         serializer = MovieSerializer(movie, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -55,6 +57,8 @@ class MovieDetailView(APIView):
 
     def delete(self, request, pk):
         movie = self.get_object(pk)
+        if movie.user != request.user:
+            return Response({"error": "You do not have permission to delete this movie."}, status=status.HTTP_403_FORBIDDEN)
         movie.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
@@ -94,7 +98,7 @@ class TagDetailView(APIView):
 
     def delete(self, request, pk):
         tag = self.get_object(pk)
-        tag.movies.clear()
+        tag.collections.clear()
         tag.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
@@ -108,9 +112,9 @@ class CollectionListView(APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        serializer = CollectionSerializer(data=request.data)
+        serializer = CollectionCreateSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
